add funcs: handle cards and user info

+++ src/components/utils/Api.js
@@ -1,3 +1,5 @@
+import userEvent from "@testing-library/user-event";
+
 //*API
 export default class Api {
   constructor({ baseUrl, headers }) {
@@ -54,18 +56,18 @@ export default class Api {
     }).then(this._check);
   }
 
-  likeCard(cardId) {
-    return fetch(`${this._url}cards/${cardId}/likes`, {
-      method: "PUT",
-      headers: this._headers,
-    }).then(this._check);
-  }
-
-  dislikeCard(cardId) {
-    return fetch(`${this._url}cards/${cardId}/likes`, {
-      method: "DELETE",
-      headers: this._headers,
-    }).then(this._check);
+  likeCard(cardId, isLiked) {
+    if (isLiked) {
+      return fetch(`${this._url}cards/${cardId}/likes`, {
+        method: "PUT",
+        headers: this._headers,
+      }).then(this._check);
+    } else {
+      return fetch(`${this._url}cards/${cardId}/likes`, {
+        method: "DELETE",
+        headers: this._headers,
+      }).then(this._check);
+    }
   }
 
   setAvatar(avatarNew) {
